third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-Id.html

<!DOCTYPE html>
<html>
<head>
<title>RTCPeerConnection.id</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
promise_test(function() {
  var pc = new RTCPeerConnection();
  assert_equals(typeof pc.id, 'string');
  return function(){};
}, 'RTCPeerConnection.id is a string.');

promise_test(function() {
  var pc1 = new RTCPeerConnection();
  var pc2 = new RTCPeerConnection();
  assert_not_equals(pc1.id, pc2.id);
  return function(){};
}, 'RTCPeerConnection.id is unique between RTCPeerConnection objects.');
</script>
</body>
</html>
